webapp/requirements.txt
starlette
jinja2
python-dotenv
aiofiles
# Required for deploying to Azure
uvicorn
uvloop
httptools